+++ hft-trade-bot/src/core/main.cpp
@@ -509,8 +509,8 @@ int main(int argc, char* argv[]) {
         uint16_t    id;
     };
     std::vector<SymbolEntry> symbol_entries;
-    for (uint16_t i = 0; i < config.symbols.size(); ++i) {
-        symbol_entries.push_back({config.symbols[i], config.symbols[i].c_str(), i});
+    for (size_t i = 0; i < config.symbols.size(); ++i) {
+        symbol_entries.push_back({config.symbols[i], config.symbols[i].c_str(), static_cast<uint16_t>(i)});
     }
 
     // Main loop
